fix(updates): one repo identity, lone-edit forks, no empty discard [no-changelog]
Repository identity folds every GitHub spelling to owner/repo so one
package never splits into two. A fork is offered only when exactly one
rendering was edited, since it keeps one and drops the rest; edits in
several tools are sent to the package page. Use new version is withheld
when the source no longer carries the package and a discard would leave
the edited copy as it is. Updates-page copy and the customized-place
tests move to their own files to stay under the line caps.
