+++ ui/src/lib/copy.ts
@@ -177,27 +177,6 @@ export const UPDATES_UNCHECKED_TITLE = "Couldn't be checked";
 export const REMOVED_UPSTREAM_TAG = "No longer in its source";
 export const UPDATE_ALL_LABEL = "Update all";
 export const CHECK_FOR_UPDATES_LABEL = "Check for updates";
-export const FOLLOW_SOURCE_COLUMN = "Follow source";
-export const FOLLOW_SOURCE_HELP =
-  "Following packages come current when their place refreshes or you press Update, which also brings that place's other followers current. Held ones wait until you choose.";
-export const heldInLabel = (held: number, total: number): string =>
-  `Held in ${held} of ${total}`;
-export const followSourceLabel = (name: string, place: string): string =>
-  `Follow the source for ${name} in ${place}`;
-export const UPDATES_NAME_COLUMN = "Package";
-export const UPDATES_TYPE_COLUMN = "Type";
-export const UPDATES_PLACE_COLUMN = "Where";
-export const UPDATES_VERSION_COLUMN = "Version";
-export const USER_LEVEL_PLACE = "User level";
-export const placesLabel = (count: number): string =>
-  count === 1 ? "1 place" : `${count} places`;
-export const updatesSubtitle = (packages: number, places: number): string =>
-  `${packages === 1 ? "1 update" : `${packages} updates`} across ${placesLabel(places)}`;
-export const UPDATE_PACKAGE_EVERYWHERE_LABEL = "Update all";
-export const CUSTOMIZED_HERE_LABEL = "Customized here";
-export const USE_NEW_VERSION_LABEL = "Use new version…";
-export const UNFORKABLE_EDIT_NOTE =
-  "Edited in a tool whose copy can't be kept as your own";
 export const IGNORE_UPDATES_LABEL = "Stop notifying…";
 export const ignoreConfirmTitle = (name: string): string =>
   `Stop notifying about ${name}?`;
@@ -211,18 +190,7 @@ export const PINNED_UPDATE_TAG = "Held";
 export const EDITED_UPDATE_TAG = "Edited by you";
 export const UPDATE_ERROR_TITLE = "Couldn't update";
 export const updatedToastLabel = (name: string): string => `Updated ${name}`;
-export const updatedWithPlaceToastLabel = (
-  name: string,
-  place: string,
-): string => `Updated ${name} and everything else following in ${place}`;
 export const UPDATED_ALL_TOAST = "Everything is up to date";
-export const updatedSomeToastLabel = (
-  updated: number,
-  skipped: number,
-): string =>
-  `Updated ${updated === 1 ? "1 package" : `${updated} packages`} — ${skipped === 1 ? "1 customized place needs" : `${skipped} customized places need`} a decision`;
-export const nothingToUpdateToastLabel = (skipped: number): string =>
-  `Nothing to update — ${skipped === 1 ? "1 customized place needs" : `${skipped} customized places need`} a decision first`;
 
 // Fork: what happens when the app finds files you edited by hand.
 export const FORKED_BADGE_LABEL = "Forked";
