+++ ui/src/stores/updates-bulk.test.ts
@@ -39,6 +39,7 @@ function row(overrides: Partial<UpdateRow>): UpdateRow {
     blockedByLocalEdit: false,
     editedHarnesses: [],
     forkableHarness: null,
+    canDiscard: true,
     forked: false,
     mixed: false,
     removedUpstream: false,
