+++ ui/src/stores/updates.test.ts
@@ -40,6 +40,7 @@ function row(overrides: Partial<UpdateRow>): UpdateRow {
     blockedByLocalEdit: false,
     editedHarnesses: [],
     forkableHarness: null,
+    canDiscard: true,
     forked: false,
     mixed: false,
     removedUpstream: false,
